# src/main/java/com/chukchuk/haksa/global/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "척척학사 API Definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/target-gpa": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "목표 GPA 설정",
        "description": "로그인된 사용자의 목표 GPA를 저장합니다.",
        "operationId": "setTargetGpa",
        "parameters": [
          {
            "name": "targetGpa",
            "in": "query",
            "description": "목표 GPA",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 3.8
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/get-academic": {
      "get": {
        "tags": [
          "Academic Record"
        ],
        "summary": "학기별 성적 및 수강 과목 정보 조회",
        "description": "지정한 학기(year, semester)에 해당하는 성적 및 수강 과목 정보를 조회합니다.",
        "operationId": "getAcademicRecord",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "description": "연도",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 2024
          },
          {
            "name": "semester",
            "in": "query",
            "description": "학기",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": "10, 15, 20 ..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicRecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SemesterGradeDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "description": "이수 연도",
            "format": "int32",
            "example": 2024
          },
          "semester": {
            "type": "integer",
            "description": "학기 코드 (10: 1학기, 15: 여름학기, 20: 2학기, 25: 겨울학기)",
            "format": "int32",
            "example": 10
          },
          "earnedCredits": {
            "type": "integer",
            "description": "취득 학점",
            "format": "int32",
            "example": 15
          },
          "attemptedCredits": {
            "type": "integer",
            "description": "신청 학점",
            "format": "int32",
            "example": 18
          },
          "semesterGpa": {
            "type": "number",
            "description": "학기 GPA (평점 평균)",
            "example": 3.85
          },
          "classRank": {
            "type": "integer",
            "description": "석차",
            "format": "int32",
            "nullable": true,
            "example": 5
          },
          "totalStudents": {
            "type": "integer",
            "description": "전체 학생 수",
            "format": "int32",
            "nullable": true,
            "example": 150
          },
          "percentile": {
            "type": "number",
            "description": "백분율",
            "example": 92.4
          }
        },
        "description": "학기 성적 정보"
      },
...
      "AcademicRecordResponse": {
        "type": "object",
        "properties": {
          "semesterGrade": {
            "$ref": "#/components/schemas/SemesterGradeDto"
          },
          "courses": {
            "$ref": "#/components/schemas/Courses"
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          }
        },
        "description": "학기별 성적 및 수강 과목 응답"
      },
      "CourseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "수강 ID"
          },
          "courseName": {
            "type": "string",
            "description": "과목명"
          },
          "courseCode": {
            "type": "string",
            "description": "학수번호"
          },
          "areaType": {
            "type": "string",
            "description": "영역 (전공/교양 등)"
          },
          "credits": {
            "type": "integer",
            "description": "학점",
            "format": "int32"
          },
          "professor": {
            "type": "string",
            "description": "교수명"
          },
          "grade": {
            "type": "string",
            "description": "성적"
          },
          "score": {
            "type": "integer",
            "description": "실 점수",
            "format": "int32"
          },
          "isRetake": {
            "type": "boolean",
            "description": "재수강 여부"
          },
          "isOnline": {
            "type": "boolean",
            "description": "사이버 강의 여부"
          },
          "year": {
            "type": "integer",
            "description": "이수 연도",
            "format": "int32"
          },
          "semester": {
            "type": "integer",
            "description": "이수 학기",
            "format": "int32"
          },
          "originalScore": {
            "type": "integer",
            "description": "원점수",
            "format": "int32"
          }
        },
        "description": "수강 과목 상세 정보"
      },
      "Courses": {
        "type": "object",
        "properties": {
          "major": {
            "type": "array",
            "description": "전공 과목 목록",
            "items": {
              "$ref": "#/components/schemas/CourseDetailDto"
            }
          },
          "liberal": {
            "type": "array",
            "description": "교양 과목 목록",
            "items": {
              "$ref": "#/components/schemas/CourseDetailDto"
            }
          }
        },
        "description": "수강 과목 목록"
      },
      "Summary": {
        "type": "object",
        "properties": {
          "totalEarnedCredits": {
            "type": "integer",
            "description": "총 취득 학점",
            "format": "int32",
            "example": 92
          },
          "cumulativeGpa": {
            "type": "number",
            "description": "누적 GPA",
            "example": 3.75
          },
          "majorGpa": {
            "type": "number",
            "description": "전공 GPA",
            "format": "double",
            "example": 3.85
          },
          "percentile": {
            "type": "number",
            "description": "백분위",
            "example": 88.5
          }
        },
        "description": "학업 요약 정보"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 인증 토큰을 입력하세요"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}